requirements.txt
fastapi[all]
uvicorn
sqlalchemy
alembic
pydantic
pytest
httpx
requests
python-jose[cryptography]
passlib[bcrypt]
psycopg2-binary 